+++ java/src/main/java/io/github/nicklambourne/slackblocks/element/ChannelMultiSelectElement.java
@@ -198,6 +198,8 @@ public final class ChannelMultiSelectElement extends BlockElement implements Inp
     /**
      * Sets the maximum number of items a user can select.
      *
+     * <p>Must be at least 1.
+     *
      * @param value value for Slack's {@code max_selected_items} field
      * @return this builder
      */