+++ java/src/main/java/io/github/nicklambourne/slackblocks/payload/ModalView.java
@@ -338,6 +338,8 @@ public final class ModalView implements SlackObject {
     /**
      * Sets a workspace-unique identifier you can use to update the view later.
      *
+     * <p>Slack allows at most 255 characters.
+     *
      * @param value value for Slack's {@code external_id} field
      * @return this builder
      */